+++ sumologic/resource_sumologic_partition.go
@@ -67,6 +67,12 @@ func resourceSumologicPartition() *schema.Resource {
 				Type:     schema.TypeBool,
 				Optional: true,
 			},
+			"is_included_in_default_search": {
+				Type:        schema.TypeBool,
+				Description: "Indicates whether the partition is included in the default search scope. Configuring this property is exclusively permitted for flex partitions.",
+				Optional:    true,
+				Default:     true,
+			},
 		},
 	}
 }
@@ -84,7 +90,7 @@ func resourceSumologicPartitionCreate(d *schema.ResourceData, meta interface{}) 
 		d.SetId(createdSpartition.ID)
 	}
 
-	return resourceSumologicPartitionUpdate(d, meta)
+	return resourceSumologicPartitionRead(d, meta)
 }
 
 func resourceSumologicPartitionRead(d *schema.ResourceData, meta interface{}) error {
@@ -113,6 +119,7 @@ func resourceSumologicPartitionRead(d *schema.ResourceData, meta interface{}) er
 	d.Set("is_active", spartition.IsActive)
 	d.Set("total_bytes", spartition.TotalBytes)
 	d.Set("index_type", spartition.IndexType)
+	d.Set("is_included_in_default_search", spartition.IsIncludedInDefaultSearch)
 
 	return nil
 }
@@ -134,11 +141,23 @@ func resourceSumologicPartitionUpdate(d *schema.ResourceData, meta interface{}) 
 }
 
 func resourceToPartition(d *schema.ResourceData) Partition {
+
+	analyticsTier := d.Get("analytics_tier").(string)
+	isIncludedInDefaultSearch := d.Get("is_included_in_default_search").(bool)
+
+	var isIncludedInDefaultSearchPtr *bool
+	if analyticsTier == "flex" || analyticsTier == "" {
+		isIncludedInDefaultSearchPtr = new(bool)
+		*isIncludedInDefaultSearchPtr = isIncludedInDefaultSearch
+	} else {
+		isIncludedInDefaultSearchPtr = nil
+	}
+
 	return Partition{
 		ID:                               d.Id(),
 		Name:                             d.Get("name").(string),
 		RoutingExpression:                d.Get("routing_expression").(string),
-		AnalyticsTier:                    d.Get("analytics_tier").(string),
+		AnalyticsTier:                    analyticsTier,
 		RetentionPeriod:                  d.Get("retention_period").(int),
 		IsCompliant:                      d.Get("is_compliant").(bool),
 		DataForwardingId:                 d.Get("data_forwarding_id").(string),
@@ -146,5 +165,6 @@ func resourceToPartition(d *schema.ResourceData) Partition {
 		TotalBytes:                       d.Get("total_bytes").(int),
 		IndexType:                        d.Get("index_type").(string),
 		ReduceRetentionPeriodImmediately: d.Get("reduce_retention_period_immediately").(bool),
+		IsIncludedInDefaultSearch:        isIncludedInDefaultSearchPtr,
 	}
 }